requirements.txt
accelerate==0.19.0
aiohttp==3.8.4
async-lru==2.0.2
bs4==0.0.1
diffusers==0.16.1
fqdn==1.5.1
google-search-results==2.4.2
isoduration==20.11.0
jsonpointer==2.3
jupyter==1.0.0
jupyterlab-lsp==4.2.0
lightning==2.0.2
pandas==2.0.1
pip-chill==1.0.3
safetensors==0.3.1
scipy==1.10.1
torchvision==0.15.2
transformers==4.29.2
uri-template==1.2.0
wandb==0.15.3
webcolors==1.13
